projects/bmt_eco_scheduler-project/bin/s02_08_poetry_lock.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-

try:
    from bmt_eco_scheduler.ops import poetry_lock

    poetry_lock()
except ImportError:
    from automation.api import pyproject_ops

    pyproject_ops.poetry_lock()
